Position coordinates in double

+++ src/objects/LoaderParams.hpp
@@ -5,7 +5,7 @@
 
 class LoaderParams {
     public:
-        LoaderParams(int x, int y, int width, int height, std::string
+        LoaderParams(double x, double y, int width, int height, std::string
                 textureID, int callbackID = 0, int numFrames = 1, int animSpeed = 0) :
             m_x(x),
             m_y(y),
@@ -15,8 +15,8 @@ class LoaderParams {
             m_numFrames(numFrames),
             m_callbackID(callbackID),
             m_animSpeed(animSpeed) {}
-        int getX() const { return m_x; }
-        int getY() const { return m_y; }
+        double getX() const { return m_x; }
+        double getY() const { return m_y; }
         int getWidth() const { return m_width; }
         int getHeight() const { return m_height; }
         int getNumFrames() const { return m_numFrames; }
@@ -25,8 +25,8 @@ class LoaderParams {
         std::string getTextureID() const { return m_textureID; }
 
     private:
-        int m_x;
-        int m_y;
+        double m_x;
+        double m_y;
         int m_width;
         int m_height;
         std::string m_textureID;
